Flask-4-7/requirements.txt
flask
sqlmodel
bcrypt
numpy<2.0.0
onnxruntime 
opencv-python-headless
python-dateutil
humanize==0.2
psycopg2-binary 